Add a separate task for running scapegoat analysis. Fixes #19
Two things to know about this patch. First, and most importantly, it forces the
plugin to depend on sbt 0.13.8 due to the use of `Def.sequential` to control
task ordering in the `scapegoat` task. Secondly, and the reason that we need to
use Def.sequential, is that it's a little ugly since it depends on side effects
on the ScapegoatPlugin object to enable and disable the analysis. I tried
several different ways to do this that wouldn't have needed side effects, but
controlling the compile settings the way we need to is hard to do, since the
compile task generator depends on (at least one) experimental, private setting
key, and just copying the compile task doesn't make it pick up a
`scalacOptions` task in a different scope.