Update README.md
I had to hunt around in the source code for 30 minutes to find the import statement necessary to get the scapegoat settings in scope.  I added it to the documentation.  If this isn't a good place for it, feel free to nuke it, but it would be really helpful to have it in the docs.  Thanks.  This library rules.