Add the compiler plugin dependency to a custom Ivy configuration
Currently the `scalac-scapegoat-plugin_$scalaVersion` dependency is
added to the built-in `provided` configuration. That means the
dependency leaks into the `pom.xml` of any project that uses
sbt-scapegoat:

```
<dependency>
    <groupId>com.sksamuel.scapegoat</groupId>
    <artifactId>scalac-scapegoat-plugin_2.13.1</artifactId>
    <version>1.4.1</version>
    <scope>provided</scope>
</dependency>
```

This PR avoids that by creating a new private Ivy configuration called
`scapegoat` and adding the dependency to that configuration. It still
leaks into the Ivy XML but not into the pom. Given that most people
publish their libraries using Maven, I think that's a reasonable
improvement.